Bugfix: Wrong case in directory name
The directory name 'iamstuartwilson' did not match the PSR-0 namespace
name 'Iamstuartwilson', so that the Composer autoloader was not able
to include the StravaApi class.

The error appeared only on case-sensitive file systems (e.g. ext4 or XFS)
but not on case-insensitive file systems (e.g. HFS+) so debugging took
some minutes... ;-)