add methods for accessing the HTTP response headers
It's possible to access the HTTP response headers with two added methods:

- getResponseHeaders()
- getResponseHeader($header)

The first one returns all HTTP headers as an array while the second returns
the header value for the given header name (and throws an exception if the
header name does not exist).

The existing public API of the StravaAPI class is unchanged, only two new
public methods were introduced.